# Replace abstract Integer with concrete Int type