Resolves https://github.com/JuliaFEM/AbaqusReader.jl/issues/50

Also change AbstractString to String in model dictionary.

This may improve performance, see:
https://docs.julialang.org/en/v1/manual/performance-tips/#Avoid-containers-with-abstract-type-parameters-1

Benchmarks
==========

It seems there is a very small improvement in performance,
but allocations and memory have been reduced.

```
julia> import Logging; Logging.disable_logging(Logging.Error)
julia> using AbaqusReader, BenchmarkTools
julia> @benchmark abaqus_read_mesh("tet4.inp")
```

Before:
```
BenchmarkTools.Trial:
  memory estimate:  980.23 KiB
  allocs estimate:  17939
  --------------
  minimum time:     1.050 ms (0.00% GC)
  median time:      1.140 ms (0.00% GC)
  mean time:        1.223 ms (6.28% GC)
  maximum time:     5.272 ms (77.52% GC)
  --------------
  samples:          4082
  evals/sample:     1
```

After:
```
BenchmarkTools.Trial:
  memory estimate:  916.25 KiB
  allocs estimate:  15570
  --------------
  minimum time:     1.004 ms (0.00% GC)
  median time:      1.094 ms (0.00% GC)
  mean time:        1.174 ms (6.58% GC)
  maximum time:     5.905 ms (78.07% GC)
  --------------
  samples:          4251
  evals/sample:     1
```