@TeroFrondelius 2020-02-25
The build fails: https://travis-ci.org/JuliaFEM/AbaqusReader.jl/builds/654216321?utm_source=github_status&utm_medium=notification

There seems to be three broken tests.