@benzwick 2020-02-25
> There seems to be three broken tests.

This should be fixed now in #60